requirements.txt
numpy
nltk
pyarrow
requests
#inference on cpu
tensorflow-cpu
#uncomment bellow and comment above to train on a GPU
#tensorflow>=2.2.0
keras
